Revamped vision to use yoloV11
