+++ arachne_control.py
@@ -45,7 +45,6 @@ class ArachneController:
         if self.j.get_init():
             print("Joystick Ready!")
         self.autonomous = False  # teleop mode for default
-        # self.vision = Vision(self)
 
         # Initializing all leg ids
         self.all_legs = (2, 3, 4, 5, 1, 0)
@@ -76,7 +75,7 @@ class ArachneController:
             # Left Side
 
             [704, 1904, -1, 1468.25, 90, 832.0, 999],  # 9
-            [496, 1216, -1, 807, 45, 528, 999],  # 10. #FIXME: replaced Servo, needs new Cal
+            [496, 1216, 1, 807, 45, 1216, 999],  # 10. #FIXME: replaced Servo, needs new Cal
             [1248, 2080, 1, 1450.00, 45, 2031, 999],  # 11   # was x, x, 1550
 
             [608, 1760, -1, 1615, 90, 784.0, 999],  # 12
@@ -216,7 +215,7 @@ class ArachneController:
                         self.autonomous = not self.autonomous
                         if self.autonomous and not autonomousThread.is_alive():
                             autonomousThread.start()
-                        else:
+                        elif not self.autonomous:
                             data = {"point": "BREAK"} # end autonomous thread
                             autonomousThread.join()
                         status = "on" if self.autonomous else "off"
@@ -392,8 +391,9 @@ class ArachneController:
     @staticmethod
     def controller_mixin(pt):  # points is a String to be parsed
         controller = ArachneController(debug=False)
-        while pt[0] != "BREAK":
-            # print("IN WHILE LOOP")
+        while True:
+            if pt[0] == "BREAK":
+                break
             if pt[0] == "NA":  # scanning for an object
                 print("TURN")
                 controller.crab_walk_turn(10)
@@ -401,113 +401,3 @@ class ArachneController:
                 print("CRAB WALK")
                 controller.crab_walk_2(0, 30, 1)
 
-
-# File courtesy of messing around with chatgpt
-class Vision:
-    def __init__(self, ac: ArachneController):
-        self.ac = ac
-        # Paths
-        self.MODEL_PATH = "vision/model/coco_ssd_mobilenet_v1_1.0_quant_2018_06_29.tflite"
-        self.LABEL_PATH = "vision/model/labelmap.txt"  # COCO Dataset
-
-        # Load labels
-        with open(self.LABEL_PATH, 'r') as f:
-            self.labels = [line.strip() for line in f.readlines()]
-
-        # Load model
-        self.interpreter = Interpreter(model_path=self.MODEL_PATH)
-        self.interpreter.allocate_tensors()
-
-        self.input_details = self.interpreter.get_input_details()
-        self.output_details = self.interpreter.get_output_details()
-
-        input_shape = self.input_details[0]['shape']
-        self.height = input_shape[1]
-        self.width = input_shape[2]
-        self.floating_model = self.input_details[0]['dtype'] == np.float32
-
-        self.new_frame_time = 0
-        self.prev_frame_time = 0
-        self.cap = cv.VideoCapture(0)
-
-    def tick(self) -> bool:  # 1 frame
-
-        retrieved, frame = self.cap.read()
-
-        if not retrieved:
-            print("Stream has likely ended")
-            return False
-
-        # cv.imshow("stream", frame)
-        # https://stackoverflow.com/questions/5217519/what-does-opencvs-cvwaitkey-function-do <-- how waitKey works
-
-        image_rgb = cv.cvtColor(frame, cv.COLOR_BGR2RGB)
-        image_resized = cv.resize(image_rgb, (self.width, self.height))
-        input_data = np.expand_dims(image_resized, axis=0)
-
-        if self.floating_model:
-            input_data = (np.float32(input_data) - 127.5) / 127.5
-        else:
-            input_data = np.uint8(input_data)
-
-        # Run inference
-        self.interpreter.set_tensor(self.input_details[0]['index'], input_data)
-        self.interpreter.invoke()
-
-        # Get outputs
-        boxes = self.interpreter.get_tensor(self.output_details[0]['index'])[0]
-        classes = self.interpreter.get_tensor(self.output_details[1]['index'])[0]
-        scores = self.interpreter.get_tensor(self.output_details[2]['index'])[0]
-        num = self.interpreter.get_tensor(self.output_details[3]['index'])[0]
-
-        # Draw detections
-        imH, imW, _ = frame.shape
-        detected = False
-        for i in range(int(num)):
-            if scores[i] > 0.5:
-                detected = True
-                ymin = int(max(1, boxes[i][0] * imH))
-                xmin = int(max(1, boxes[i][1] * imW))
-                ymax = int(min(imH, boxes[i][2] * imH))
-                xmax = int(min(imW, boxes[i][3] * imW))
-
-                class_id = int(classes[i])
-                label = self.labels[class_id] if class_id < len(self.labels) else "N/A"
-                confidence = int(scores[i] * 100)
-
-                center = (xmin + int(abs(xmin - xmax) / 2), ymin + int(abs(ymax - ymin) / 2))
-                #cv.rectangle(frame, (xmin, ymin), (xmax, ymax), (0, 255, 0), 2)
-                #cv.putText(frame, f"{label} ({confidence}%)", (xmin, ymin - 10),
-                # cv.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
-                # Show output
-
-        self.new_frame_time = time.time()
-        fps = 1 / (self.new_frame_time - self.prev_frame_time)
-        self.prev_frame_time = self.new_frame_time  # Get fps
-
-        # Display FPS on the frame (optional)
-        # cv.putText(frame, f"FPS: {fps}", (10, 30), cv.FONT_HERSHEY_SIMPLEX, 1, (0, 255, 0), 2)
-        # cv.imshow("Object Detection", frame)
-
-        print(num)
-        if not detected:  # scanning
-            self.ac.crab_walk_turn(10)  # keep turning until you detect something
-        # # else: # start walking forward
-        # #     cv.line(frame, center, center, (0, 0, 0), 20) # if you start walking forward, put a dot on the center of the detected object
-        # #     # fine adjustment
-        # #     actual_w = cv.CAP_PROP_FRAME_WIDTH/2
-        # #     if 0 <= (center[0]-actual_w) <= 10:
-        # #         arachne_control.crab_walk_turn(30) # move left
-        # #     elif 0<= (actual_w - center[0]) <= 10:
-        # #         arachne_control.crab_walk_turn(-30) # move right
-
-        # #     arachne_control.crab_walk_2(0, 30, 1)
-
-        if cv.waitKey(1) == ord("q"):  # ESC to quit
-            return False
-
-        return True
-
-        # # if joystick.get_button(1): # exit program if the toggle is pressed
-        # #     self.cap.release()
-        # #     cv.destroyAllWindows()
